change datatype to match database

+++ Core/Modules/Orders/Models/Payment.cs
@@ -5,7 +5,7 @@ namespace Pharmacy.Domain.Modules.Orders.Models;
 
 public sealed class Payment : BaseModel<int>
 {
-    public decimal Paid {get; set;} = 0;
+    public double Paid {get; set;} = 0;
     public DateTime CreatedAt {get;} = DateTime.UtcNow;
     public required Guid CustomerId {get; set;}
 }
